requirements.txt
timm==0.6.13
transformers
fairscale
pycocoevalcap
opencv-python
qrcode[pil]
pytorch_lightning
kornia
pydantic
segment_anything
omegaconf
openai>=1.23.3
anthropic